cleaning a bit
